+++ python/GafferSceneTest/ShaderTweaksTest.py
@@ -184,7 +184,7 @@ class ShaderTweaksTest( GafferSceneTest.SceneTestCase ) :
 
 		tweakedNetwork = tweaks["out"].attributes( "/plane" )["surface"]
 		self.assertEqual( len( tweakedNetwork ), 2 )
-		self.assertEqual( tweakedNetwork.input( ( "surface", "c" ) ), ( "texture", "" ) )
+		self.assertEqual( tweakedNetwork.input( ( "surface", "c" ) ), ( "texture", "out" ) )
 
 		tweakedNetwork.removeShader( "texture" )
 		self.assertEqual( tweakedNetwork, originalNetwork )
@@ -257,7 +257,7 @@ class ShaderTweaksTest( GafferSceneTest.SceneTestCase ) :
 
 		originalNetwork = assignment["out"].attributes( "/plane" )["surface"]
 		self.assertEqual( len( originalNetwork ), 2 )
-		self.assertEqual( originalNetwork.input( ( "surface", "c" ) ), ( "texture1", "" ) )
+		self.assertEqual( originalNetwork.input( ( "surface", "c" ) ), ( "texture1", "out" ) )
 
 		textureShader2 = GafferSceneTest.TestShader( "texture2" )
 
@@ -271,7 +271,7 @@ class ShaderTweaksTest( GafferSceneTest.SceneTestCase ) :
 
 		tweakedNetwork = tweaks["out"].attributes( "/plane" )["surface"]
 		self.assertEqual( len( tweakedNetwork ), 2 )
-		self.assertEqual( tweakedNetwork.input( ( "surface", "c" ) ), ( "texture2", "" ) )
+		self.assertEqual( tweakedNetwork.input( ( "surface", "c" ) ), ( "texture2", "out" ) )
 
 		textureShader2["enabled"].setValue( False )
 		tweakedNetwork = tweaks["out"].attributes( "/plane" )["surface"]